Fix: Rail/road type conversion data is not populated on new game.
This means that the rail/road type conversion that happens if NewGRFs are changed does not work the first time for a new game, only with a savegame.
