+++ src/newgrf_railtype.cpp
@@ -14,6 +14,7 @@
 #include "timer/timer_game_calendar.h"
 #include "depot_base.h"
 #include "town.h"
+#include "tunnelbridge_map.h"
 
 #include "safeguards.h"
 
@@ -173,4 +174,74 @@ uint8_t GetReverseRailTypeTranslation(RailType railtype, const GRFFile *grffile)
 
 	/* If not found, return as invalid */
 	return 0xFF;
+}
+
+std::vector<LabelObject<RailTypeLabel>> _railtype_list;
+
+/**
+ * Test if any saved rail type labels are different to the currently loaded
+ * rail types. Rail types stored in the map will be converted if necessary.
+ */
+void ConvertRailTypes()
+{
+	std::vector<RailType> railtype_conversion_map;
+	bool needs_conversion = false;
+
+	for (auto it = std::begin(_railtype_list); it != std::end(_railtype_list); ++it) {
+		RailType rt = GetRailTypeByLabel(it->label);
+		if (rt == INVALID_RAILTYPE) {
+			rt = RAILTYPE_RAIL;
+		}
+
+		railtype_conversion_map.push_back(rt);
+
+		/* Conversion is needed if the rail type is in a different position than the list. */
+		if (it->label != 0 && rt != std::distance(std::begin(_railtype_list), it)) needs_conversion = true;
+	}
+
+	if (!needs_conversion) return;
+
+	for (const auto t : Map::Iterate()) {
+		switch (GetTileType(t)) {
+			case MP_RAILWAY:
+				SetRailType(t, railtype_conversion_map[GetRailType(t)]);
+				break;
+
+			case MP_ROAD:
+				if (IsLevelCrossing(t)) {
+					SetRailType(t, railtype_conversion_map[GetRailType(t)]);
+				}
+				break;
+
+			case MP_STATION:
+				if (HasStationRail(t)) {
+					SetRailType(t, railtype_conversion_map[GetRailType(t)]);
+				}
+				break;
+
+			case MP_TUNNELBRIDGE:
+				if (GetTunnelBridgeTransportType(t) == TRANSPORT_RAIL) {
+					SetRailType(t, railtype_conversion_map[GetRailType(t)]);
+				}
+				break;
+
+			default:
+				break;
+		}
+	}
+}
+
+/** Populate railtype label list with current values. */
+void SetCurrentRailTypeLabelList()
+{
+	_railtype_list.clear();
+
+	for (RailType rt = RAILTYPE_BEGIN; rt != RAILTYPE_END; rt++) {
+		_railtype_list.push_back({GetRailTypeInfo(rt)->label, 0});
+	}
+}
+
+void ClearRailTypeLabelList()
+{
+	_railtype_list.clear();
 }
