+++ src/newgrf_railtype.h
@@ -59,4 +59,8 @@ SpriteID GetCustomSignalSprite(const RailTypeInfo *rti, TileIndex tile, SignalTy
 RailType GetRailTypeTranslation(uint8_t railtype, const GRFFile *grffile);
 uint8_t GetReverseRailTypeTranslation(RailType railtype, const GRFFile *grffile);
 
+void ConvertRailTypes();
+void SetCurrentRailTypeLabelList();
+void ClearRailTypeLabelList();
+
 #endif /* NEWGRF_RAILTYPE_H */
