+++ src/newgrf_roadtype.cpp
@@ -14,6 +14,7 @@
 #include "timer/timer_game_calendar.h"
 #include "depot_base.h"
 #include "town.h"
+#include "tunnelbridge_map.h"
 
 #include "safeguards.h"
 
@@ -166,4 +167,68 @@ uint8_t GetReverseRoadTypeTranslation(RoadType roadtype, const GRFFile *grffile)
 
 	/* If not found, return as invalid */
 	return 0xFF;
+}
+
+std::vector<LabelObject<RoadTypeLabel>> _roadtype_list;
+
+/**
+ * Test if any saved road type labels are different to the currently loaded
+ * road types. Road types stored in the map will be converted if necessary.
+ */
+void ConvertRoadTypes()
+{
+	std::vector<RoadType> roadtype_conversion_map;
+	bool needs_conversion = false;
+	for (auto it = std::begin(_roadtype_list); it != std::end(_roadtype_list); ++it) {
+		RoadType rt = GetRoadTypeByLabel(it->label);
+		if (rt == INVALID_ROADTYPE || GetRoadTramType(rt) != it->subtype) {
+			rt = it->subtype ? ROADTYPE_TRAM : ROADTYPE_ROAD;
+		}
+
+		roadtype_conversion_map.push_back(rt);
+
+		/* Conversion is needed if the road type is in a different position than the list. */
+		if (it->label != 0 && rt != std::distance(std::begin(_roadtype_list), it)) needs_conversion = true;
+	}
+	if (!needs_conversion) return;
+
+	for (TileIndex t : Map::Iterate()) {
+		switch (GetTileType(t)) {
+			case MP_ROAD:
+				if (RoadType rt = GetRoadTypeRoad(t); rt != INVALID_ROADTYPE) SetRoadTypeRoad(t, roadtype_conversion_map[rt]);
+				if (RoadType rt = GetRoadTypeTram(t); rt != INVALID_ROADTYPE) SetRoadTypeTram(t, roadtype_conversion_map[rt]);
+				break;
+
+			case MP_STATION:
+				if (IsStationRoadStop(t) || IsRoadWaypoint(t)) {
+					if (RoadType rt = GetRoadTypeRoad(t); rt != INVALID_ROADTYPE) SetRoadTypeRoad(t, roadtype_conversion_map[rt]);
+					if (RoadType rt = GetRoadTypeTram(t); rt != INVALID_ROADTYPE) SetRoadTypeTram(t, roadtype_conversion_map[rt]);
+				}
+				break;
+
+			case MP_TUNNELBRIDGE:
+				if (GetTunnelBridgeTransportType(t) == TRANSPORT_ROAD) {
+					if (RoadType rt = GetRoadTypeRoad(t); rt != INVALID_ROADTYPE) SetRoadTypeRoad(t, roadtype_conversion_map[rt]);
+					if (RoadType rt = GetRoadTypeTram(t); rt != INVALID_ROADTYPE) SetRoadTypeTram(t, roadtype_conversion_map[rt]);
+				}
+				break;
+
+			default:
+				break;
+		}
+	}
+}
+
+/** Populate road type label list with current values. */
+void SetCurrentRoadTypeLabelList()
+{
+	_roadtype_list.clear();
+	for (RoadType rt = ROADTYPE_BEGIN; rt != ROADTYPE_END; rt++) {
+		_roadtype_list.push_back({GetRoadTypeInfo(rt)->label, GetRoadTramType(rt)});
+	}
+}
+
+void ClearRoadTypeLabelList()
+{
+	_roadtype_list.clear();
 }
