+++ src/saveload/saveload_internal.h
@@ -29,7 +29,6 @@ void AfterLoadVehiclesPhase2(bool part_of_load);
 void FixupTrainLengths();
 void AfterLoadStations();
 void AfterLoadRoadStops();
-void ResetLabelMaps();
 void AfterLoadLabelMaps();
 void AfterLoadStoryBook();
 void AfterLoadLinkGraphs();
